IDA6-99: fixed thecode standard issue move add-patient types to types folder
